Move data structure hooks into @react-stately/data
