+++ packages/@react-stately/data/README.md
@@ -1,0 +1,3 @@
+# @react-stately/data
+
+This package is part of [react-spectrum](https://github.com/adobe-private/react-spectrum-v3). See the repo for more details.
